+++ src/main/java/vn/coursemanage/gui/course/OnsiteCourseManagerGUI.java
@@ -471,7 +471,7 @@ public class OnsiteCourseManagerGUI extends javax.swing.JPanel {
         searchMap.add(new SearchByFields(txtLocation.getText(), "location"));
         searchMap.add(new SearchByFields(((Item) cbbDepartment.getSelectedItem()).getId(), "departmentId"));
         if (!txtCredits.getText().equals("")) {
-            searchMap.add(new SearchByFields(Long.parseLong(txtCredits.getText()), "credits"));
+            searchMap.add(new SearchByFields(Double.parseDouble(txtCredits.getText()), "credits"));
         }
         if (!txtDays.getText().equals("")) {
             searchMap.add(new SearchByFields(Integer.parseInt(txtDays.getText()), "days"));
